Move associated_item* providers to their own module
